requirements.txt
pillow
numpy
scipy
pytest
